# docs/openapi/auth.yaml
openapi: 3.0.3
info:
  title: X1 Restaurant API - Auth Module
  description: API-only authentication and customer profile management for X1
  version: 1.0.0

servers:
  - url: /api/v1
    description: API v1

paths:
  /auth/register:
    post:
      summary: Register a new customer account
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/login:
    post:
      summary: Authenticate and obtain tokens
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/token/refresh:
    post:
      summary: Rotate tokens using refresh token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refresh successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefreshResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/logout:
    post:
      summary: Revoke current tokens
      tags:
        - Authentication
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logout successful
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/email/verify/request:
    post:
      summary: Request a new verification email
      tags:
        - Email Verification
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Verification email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Email already verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/email/verify/confirm:
    post:
      summary: Confirm email with token
      tags:
        - Email Verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/password/forgot:
    post:
      summary: Request password reset email
      tags:
        - Password Reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Reset email sent (if email exists)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/password/reset:
    post:
      summary: Reset password with token
      tags:
        - Password Reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /me:
    get:
      summary: Get current user profile
      tags:
        - Profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: Update current user profile
      tags:
        - Profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /me/addresses:
    get:
      summary: List user addresses
      tags:
        - Addresses
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Addresses retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create new address
      tags:
        - Addresses
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreAddressRequest'
      responses:
        '201':
          description: Address created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /me/addresses/{id}:
    patch:
      summary: Update address
      tags:
        - Addresses
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAddressRequest'
      responses:
        '200':
          description: Address updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Address not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete address
      tags:
        - Addresses
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Address deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Address not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /me/addresses/{id}/make-default:
    post:
      summary: Set address as default
      tags:
        - Addresses
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Address set as default
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Address not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - password_confirmation
      properties:
        email:
          type: string
          format: email
          example: customer@example.com
        password:
          type: string
          minLength: 8
          example: SecurePass123!
        password_confirmation:
          type: string
          example: SecurePass123!
        first_name:
          type: string
          maxLength: 100
          example: John
        last_name:
          type: string
          maxLength: 100
          example: Doe

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: customer@example.com
        password:
          type: string
          example: SecurePass123!

    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          example: rt_xyz789...

    VerifyEmailRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          minLength: 64
          maxLength: 64

    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: customer@example.com

    ResetPasswordRequest:
      type: object
      required:
        - token
        - email
        - password
        - password_confirmation
      properties:
        token:
          type: string
          minLength: 64
          maxLength: 64
        email:
          type: string
          format: email
          example: customer@example.com
        password:
          type: string
          minLength: 8
          example: NewSecurePass456!
        password_confirmation:
          type: string
          example: NewSecurePass456!

    UpdateProfileRequest:
      type: object
      properties:
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        phone:
          type: string
          maxLength: 20
        marketing_opt_in:
          type: boolean
        birth_date:
          type: string
          format: date

    StoreAddressRequest:
      type: object
      required:
        - label
        - country
        - city
        - postal_code
        - street_line1
      properties:
        label:
          type: string
          maxLength: 50
          example: Home
        country:
          type: string
          minLength: 2
          maxLength: 2
          example: US
        city:
          type: string
          maxLength: 100
          example: New York
        postal_code:
          type: string
          maxLength: 20
          example: "10001"
        street_line1:
          type: string
          maxLength: 255
          example: 123 Main St
        street_line2:
          type: string
          maxLength: 255
          example: Apt 4B
        is_default:
          type: boolean
          default: false

    UpdateAddressRequest:
      type: object
      properties:
        label:
          type: string
          maxLength: 50
        country:
          type: string
          minLength: 2
          maxLength: 2
        city:
          type: string
          maxLength: 100
        postal_code:
          type: string
          maxLength: 20
        street_line1:
          type: string
          maxLength: 255
        street_line2:
          type: string
          maxLength: 255
        is_default:
          type: boolean

    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: customer@example.com

    CustomerProfile:
      type: object
      properties:
        first_name:
          type: string
          nullable: true
          example: John
        last_name:
          type: string
          nullable: true
          example: Doe
        phone:
          type: string
          nullable: true
          example: "+1234567890"
        marketing_opt_in:
          type: boolean
          example: false
        birth_date:
          type: string
          format: date
          nullable: true
          example: "1990-01-15"

    CustomerAddress:
      type: object
      properties:
        id:
          type: integer
          example: 1
        label:
          type: string
          example: Home
        country:
          type: string
          example: US
        city:
          type: string
          example: New York
        postal_code:
          type: string
          example: "10001"
        street_line1:
          type: string
          example: 123 Main St
        street_line2:
          type: string
          nullable: true
          example: Apt 4B
        is_default:
          type: boolean
          example: true

    AuthToken:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        access_token:
          type: string
          example: 1|abc123...
        refresh_token:
          type: string
          example: rt_xyz789...
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          example: 900

    TokenRefresh:
      type: object
      properties:
        access_token:
          type: string
          example: 2|def456...
        refresh_token:
          type: string
          example: rt_new123...
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          example: 900

    Me:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: customer@example.com
        email_verified:
          type: boolean
          example: true
        profile:
          $ref: '#/components/schemas/CustomerProfile'

    Error:
      type: object
      properties:
        title:
          type: string
          example: Validation Error
        detail:
          type: string
          example: The given data was invalid.
        status:
          type: integer
          example: 422
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email:
              - The email field is required.
        trace_id:
          type: string
          example: abc123

    AuthTokenResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/AuthToken'
        meta:
          type: object
          properties:
            email_verification_required:
              type: boolean
              example: true

    TokenRefreshResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/TokenRefresh'
        meta:
          type: object

    MessageResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            message:
              type: string
              example: Operation successful.
        meta:
          type: object

    MeResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Me'
        meta:
          type: object

    AddressResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/CustomerAddress'
        meta:
          type: object

    AddressListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CustomerAddress'
        meta:
          type: object
          properties:
            total:
              type: integer
              example: 1